pyproject.toml
[project]
name = "archipelago-bot"
version = "0.1.0"
description = "A self-hosted Discord Bot for managing Archipelago multiworld sessions."
authors = [
    {name = "Mikail Yaldiz", email = "mikail.yaldiz@gmx.de"},
]
dependencies = [
    "PyYAML>=6.0",
    "discord.py>=2.3.0",
    "python-dotenv>=0.20.0",
    "ruff>=0.13.2",
]
requires-python = ">=3.10.0"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm]
python.use_venv = true
python.venv_in_project = true

[tool.ruff]
target-version = "py310"

lint.select = ["E", "F", "I"]

lint.ignore = ["E501"]

lint.exclude = [
    ".git",
    "__pycache__",
    "build",
    "dist",
    ".venv"
]